backend/requirements.txt
flask
pymongo
werkzeug
pyjwt
flask-cors
